REFACTOR :기존 스키마파일 모듈화
