--- .env
@@ -1,9 +1,0 @@
-FLASK_APP=app.py
-FLASK_ENV=development
-FLASK_DEBUG=1
-FLASK_RUN_PORT=5001
-CLIENT_ID = "2891e05de881290af4ff8cca3fec132c"
-CLIENT_SECRET = "frV68bDw7U6IjYRvwYMiOKBxRC37ynf2"
-REDIRECT_URI = "http://localhost:5001/oauth/kakao/callback"
-SIGNOUT_REDIRECT_URI = "http://localhost:5001/oauth/kakao/logout"
-DB_ADDRESS = "postgresql://postgres:1234@localhost/postgres"
