+++ templates/test_profile.html
@@ -141,7 +141,7 @@
           body: JSON.stringify({
             query: `
                         mutation {
-                            updateProfile(
+                            updateUser(
                                 nickname: "${nickname}"
                                 birthDate: "${birthDate}"
                                 gender: "${gender}"
